Fix reconciliation and switching chains
